requirements.txt
numpy==1.22.1
Pillow==9.4.0
streamlit==1.18.1
tensorflow==2.12.0
ipykernel
altair==4.2.2
